+++ tests/provider/dwd/observation/test_parameters.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2018-2022, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
 from wetterdienst.provider.dwd.observation import (
     DwdObservationDataset,
     DwdObservationParameter,